create OfficialList
